renamed src/shared-types and moved OnlineStatusEnum to User entity file
