+++ src/server/chatgroups/chatgroups.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, UseGuards, Get, Param } from '@nestjs/common'
 import ChatGroupService from './chatgroups.service'
 import AuthGuard from '../auth/auth.guard'
-import { IChatGroupReducer } from '../../shared-types'
+import { IChatGroupReducer } from '../../types-for-both-server-and-client'
 
 @Controller('/api/chatgroup')
 @UseGuards(AuthGuard)
