--- src/shared-types/shared-enums.ts
@@ -1,4 +1,0 @@
-export enum OnlineStatusesEnum {
-  Offline = 'Offline',
-  Online = 'Online'
-}
